+++ pages/api/v2/arbitrum/pairs.ts
@@ -1,13 +1,13 @@
 import { ARBITRUM } from '@/config/constants'
-import { getPerpetualMetadata } from '@/utils/synthetics/getPerpetualMetadata'
-import { getSwapMetadata } from '@/utils/synthetics/getSwapMetadata'
+import { getPerpetualPairsInfo } from '@/utils/synthetics/getPerpetualPairsInfo'
+import { getSwapPairsInfo } from '@/utils/synthetics/getSwapPairsInfo'
 import { NextApiRequest, NextApiResponse } from 'next'
 
 async function handleRequest(_req: NextApiRequest, res: NextApiResponse) {
   const currentNetwork = ARBITRUM
   try {
-    const perpetualPairs = await getPerpetualMetadata(currentNetwork)
-    const spotPairs = await getSwapMetadata(currentNetwork)
+    const perpetualPairs = await getPerpetualPairsInfo(currentNetwork)
+    const spotPairs = await getSwapPairsInfo(currentNetwork)
 
     res.status(200).json(perpetualPairs?.concat(spotPairs ?? []))
   } catch (error) {
