+++ utils/synthetics/getPrices.ts
@@ -1,6 +1,10 @@
 import { gql } from 'graphql-request'
 import fetchGraphQL from '../../lib/fetchGraphQL'
-import { SYNTHETICS_SUBGRAPHS } from '../../config/synthetics'
+import {
+  ORACLE_KEEPER_URLS,
+  SYNTHETICS_SUBGRAPHS,
+} from '../../config/synthetics'
+import { fetchUrl } from '@/lib/fetchUrl'
 
 const query = gql`
   query TokenPrices {
@@ -46,4 +50,29 @@ export async function getContractPrices(
     console.error(e)
     return null
   }
+}
+
+type PriceInfo = {
+  tokenSymbol: string
+  high: number
+  low: number
+  open: number
+  close: number
+}
+
+export function getStableTokenPrice(tokenSymbol: string) {
+  return {
+    tokenSymbol,
+    high: 1,
+    low: 1,
+    open: 1,
+    close: 1,
+  }
+}
+
+export async function getTokensPrice(chainId: number): Promise<PriceInfo[]> {
+  const baseUrl = ORACLE_KEEPER_URLS[chainId]
+  const dailyPriceUrl = `${baseUrl}/prices/24h`
+  const pricesData: PriceInfo[] = await fetchUrl(dailyPriceUrl)
+  return pricesData
 }
